Reorganize some code.
